Create 7 Up 7 Down game
It is a gambling game in which you have to bet on whether the dice will roll 7 up (8, 9, 10, 11, 12, 13, 14) or 7 down (1, 2, 3, 4, 5, 6). If you predict it correctly you win the game and receive 1x your stake. If your prediction is wrong you lose and you don't get anything. You get 3x your stake if you predict 7 and dice rolls 7.